Reference documention for API endpoints.
This is the first part of reference documentation for API endpoints.
This commit includes relatively complete documentation for the Client
resource, and incomplete documentation for the Enrollments resource.
We will next add the remaining subresources under Enrollments, and
examples for same, then do Organizations and Projects resources.

Note that I've put this on the `feature-compass_schema` branch because
the most recent API development has been happening on that branch, in
conjunction with the implementation of the API on the Pathways COMPASS
ROSE Schema.  As per `CONTRIBUTING.md` documentation specifically
related to work done on a feature branch should be pushed to that
branch.  However, the ultimate destination for this documentation, as
for the API itself, is the `master` branch and the release branches.