Add application.properties file step to INSTALL.md
This takes care of a documentation need that we've actually had since
commit de763e60403: when `src/main/resources/application.properties`
was renamed to `application.properties.example`, that meant that the
live application's properties file wouldn't be created at install time
without some kind of extra step.  The result was that if you visited
`openhmis/api/v3/healthcheck`, you would see

    Your service is working with version

instead of, e.g. this:

    Your service is working with version 3.1.0

For now, the solution is just to add a manual step in INSTALL.md, but
some day we should automate this as part of the Maven deploy.  (Note
that this problem is separate from the problem of keeping the version
number itself up-to-date, which is discussed in issue #40.)