Add BaseDTO class, extend all DTOs, fix wrap logic
Create a common abstract class that all the DTOs extend.  This should
prevent a class of bugs that had previously been addressed in
whack-a-mole style.  Now instead of checking what the data object type
isn't (an infinite list of unknowns), we check what it is (a finite
list of knowns).

This also fixes WADL generation, which has been broken since probably
commit d0f9d1e05, when the `WrappingWriterInterceptor` was added using
the original "check what the type isn't" logic, which mis-identified
the WADL as a DTO when actually it's a byte array.