Improve exception-catching for Google authn
Although not documented to do so [1], GoogleIdToken.parse() will
apparently throw a "java.lang.IllegalArgumentException" [2] if the
token is invalid by inspection -- e.g., if it's not formatted
properly.  This commit adds a catch for that specific exception, and
furthermore adds a generic exception catch as per issue #62.

After this, going to the "/api/v3/healthcheck/authentication" endpoint
with an "Authorization" header with an invalid value will result in
the response "You are not authenticated" being sent back to the
client, instead of a 500 error.

[1] https://google.github.io/google-api-java-client/releases/1.21.0/\
    javadoc/com/google/api/client/googleapis/auth/oauth2/GoogleIdToken.html

[2] http://stackoverflow.com/questions/32489580/\
    googleauthutil-returns-accestoken-tokenverifier-expects-idtoken