Homepage edits for readability
Style paragraphs with CSS; move text into two columns; add images.  This
is still a work in progress, with more moving around and resizing to be
done.  Notably, the client app still needs to be embedded on this page.

When viewing this diff, note that a lot of the change in index.html is
whitespace only.  Try passing -w.