Reduce logging level; add new log calls: #65
Add more logging calls to the clients endpoint.  Next I'll add these to
the rest of the endpoints.

The Log4J config file determines how detailed to get in the
logging. INFO is slightly less detailed than DEBUG, so with this change
we'll get less logging output total (in both catalina.out and
OpenHMIS.log), but we will get all the INFO calls we add to the
endpoints.  Each level is inclusive of the more severe levels, so now
that we've set the logging level to INFO we'll see everything in the
INFO, WARN, ERROR, and FATAL categories, but not the lower level
categories (TRACE and DEBUG).  See
https://logging.apache.org/log4j/2.x/manual/configuration.html for more.

For future debugging we can always change the logging level in log4j.xml
to show more detail (e.g., back to DEBUG).  For now, though, using INFO
should reduce the amount of logging, as requested in #72.