Integration tests now migrate the schema
We're working to make integration tests run against a test database, and the first step is to have tests actually build the database in the test environment before the tests are run.  This does NOT populate the database with any data, but it's a nice step.  It uses Flyway (the tool we use to migrate the main database schema) which means that when we update the main schema the test schema should update without any extra steps.