Add token cache table: #68
Create a new table to store user id tokens (from Google) so that we
don't need to call out to Google every time a request comes in to the
API. In order to use the new table, create the wrapper objects as
described in DESIGN.md.

Note that this change involves a migration.  Run `mvn clean compile
flyway:migrate` to apply it to the database.