Fix token cache, store tokens: #68
Store new id tokens in the new table.  This is far from complete; see
TBDs in the code.  While id tokens are now stored correctly, they are
not yet successfully checked.

Also, add the new table to the Hibernate Session Factory, which I missed
in f7a1236 but is part of the process for adding a new table to the
database.

Note that the VARCHAR field was not large enough to actually store id
tokens, so I had to alter the migration.  If the previous version of the
migration was already applied, you can either manually alter the type of
the idToken column or, if there's no data in the table, do the
following: drop the table, delete the migration record from the
`schema_version` table, and run `flyway:migrate` again.