Check token cache; preserve current access: #68
When resolving an identity, return a user ID instead of an external
(Google) id.  Check the cache of tokens first, before making a round
trip to Google.

Note there is some hacky stuff in here to preserve the "everyone who has
auth'd to Google has READ access" rule that we've made for the demo
server.  I'd welcome comments about improving that!  See the explanation
in the code.