Change cache table, store external id: #68
As mentioned in my comment in 125171a, storing the user (internal) id in
the cache table did not work for people who were authenticated with
Google but did not exist in our users table.  Replace the internal id
with the external id in the cache table.

Note that this commit *only* changes the table and objects.  Changes to
the actual Authentication scheme are more complex, and will come in the
next commit.  That means that the code does *not* work as of this
commit.  I just wanted to break out this change from the more
substantive changes that are coming in the next commit, for ease of
reading later.