Return account object instead of id: #68
Take a token and return an account object consisting of the external id
and, if the user exists, a user object.  If the token corresponds to
someone in the Users table, the caller will have the entire user object.
If a corresponding user does not exist but the token is good, the caller
will receive the "externalId" for the token and can display that as
needed.

Note that this breaks something, so tokens are now being stored in the
cache multiple times again.  I will fix that next.

Also note comments in the code to this effect: we need to send an error
for authentication failures (bad tokens), instead of just a null account
object.  Coming soon!