Correct some endpoint paths in logging
(with @cecilia-donnelly)

We decided not to put these changes into commit effae97d because they
are unrelated to that merge.  Note that one of these changes is really
a followup to the "noncash-benefits" -> "non-cash-benefits" change
made in commit 14aeb2d4, while the other just fixes a plain old typo
that got made once and then copied-and-pasted several times.