Initial Consent Profiles
Consent profiles are a meta domain object which provides
a way to look up field access based on consents in the
database.

This commit provides an initial DAO to generate a profile
for a single client.  This can easily be reused for multiple
clients by modifying the SQL and creating a separate method.