Consent Profiles used in Client Service
Consent profiles are now looked up in the GET Client endpoint
and used to filter out any fields the current user doesn't
have access to according to the CoC and Organization
they are a part of.

NOTE: Issue #81 seems to be relevant here as well.  The organization
and CoC fields that are part of a user apparently are NOT foreign keys
and their accessors don't use the 'id' terminology.