Huge Hibernate-related changes
It turns out HQL requires a lot more explicit tying of Hibernate domain
objects.  This means that the domain object no longer exposes an 'id'
field but instead actually links to real objects.

Ultimately in order to support the more complex many-to-many aspects
of the designed sharing rules we needed to upgrade the relevant
domain objects to use that, thus enabling the complex select
found in ConsentProfilesDAO.java

The query couldn't support grouping for lower level MySQL reasons
and so this code isn't final -- the filtering to only use
the most recent flag needs to be added to the processing loop.

Furthermore, an issue needs to be created since now the newest
domain objects use a different naming and interaction convention
than the entire rest of the API.  Every domain object, DAO, and
manager should be updated to use the Hibernate standard method
of handling and navigating foreign key relationships.

Finally, this patch created a real CoC table, since it was clear
that the ProjectCoCs in the HUD standard is actually a many-to-many
link table with no associated CoC table.  In short: CoCs in our system
could have codes (Strings) but not Ids.  Now we have a real CoC table.

Sorry for the mega commit, but it represents a wormhole.