Return meaningful error instead of 500
Because of the new consent rules, users who are authenticated but not
authorized can no longer view clients.  That kind of user had been
receiving a 500 response before this change.

I don't think this is necessarily the best way to fix this problem
(really we need to handle authn'd-but-not-authz'd users in some
consistent way), but getting an access denied error is better than a
500.