Add instructions for granting consent
The default is for a client not to be shared, but in some test cases
users may want to manually allow sharing on a client.  Add an
explanation of how to do that.

Also, fix a typo that reversed argument order in a function.
`getConsentProfile()`'s argument order in its declaration now matches
the order of params used by its callers.